% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_asa24}
\alias{demo_asa24}
\title{Example Dietary Recall Data (ASA24)}
\format{
A data frame with 68 rows and 133 variables.
}
\source{
Adapted from \link{GitHub repo link here}
}
\usage{
demo_asa24
}
\description{
This dataset is a ASA24 data adapted from a public GitHub repository: https://github.com/knights-lab/dietstudy_analyses.
Ref: Johnson et al., 2019, Cell Host & Microbe 25, 789–802
It includes dietary records with 133 variables such as UserID, RecallNo,
Occ_No, Food_Description, nutrient content, and intake amounts.
}
\keyword{datasets}
