% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{re}
\alias{re}
\title{Relative efficiency (RE) calculation}
\usage{
re(od, subod, rounded = TRUE, verbose = TRUE)
}
\arguments{
\item{od}{Returned object of first design (e.g., unconstrained optimal design)
from function \code{\link{od.1}}, \code{\link{od.2}},
\code{\link{od.3}}, \code{\link{od.4}}, \code{\link{od.2m}},
\code{\link{od.3m}}, or \code{\link{od.4m}}.}

\item{subod}{Returned object of second design (e.g., constrained optimal design)
from function \code{\link{od.1}}, \code{\link{od.2}},
\code{\link{od.3}}, \code{\link{od.4}}, \code{\link{od.2m}},
\code{\link{od.3m}}, or \code{\link{od.4m}}.}

\item{rounded}{Logical; round the values of \code{p}, \code{n}/\code{J}/\code{K}
that are from functions to two decimal places and integer, respectively if TRUE,
no rounding if FALSE; default is TRUE.}

\item{verbose}{Logical; print the value of relative efficiency if TRUE,
otherwise not; default is TRUE.}
}
\value{
Relative efficiency value.
}
\description{
Calculate the relative efficiency (RE) between two designs, it returns
    same results as those from function \code{\link{rpe}}.
}
\examples{
# Unconstrained optimal design of 2-level CRT #----------
  myod1 <- od.2(icc = 0.2, r12 = 0.5, r22 = 0.5, c1 = 1, c2 = 5, c1t = 1, c2t = 50,
              varlim = c(0.01, 0.02))
# Constrained optimal design with n = 20
  myod2 <- od.2(icc = 0.2, r12 = 0.5, r22 = 0.5, c1 = 1, c2 = 5, c1t = 1, c2t = 50,
              n = 20, varlim = c(0.005, 0.025))
# Relative efficiency (RE)
  myre <- re(od = myod1, subod= myod2)
  myre$re # RE = 0.88
# Constrained optimal design with p = 0.5
  myod2 <- od.2(icc = 0.2, r12 = 0.5, r22 = 0.5, c1 = 1, c2 = 5, c1t = 1, c2t = 50,
             p = 0.5, varlim = c(0.005, 0.025))
# Relative efficiency (RE)
  myre <- re(od = myod1, subod= myod2)
  myre$re # RE = 0.90

# Unconstrained optimal design of 3-level CRT #----------
  myod1 <- od.3(icc2 = 0.2, icc3 = 0.1, r12 = 0.5, r22 = 0.5, r32 = 0.5,
             c1 = 1, c2 = 5, c3 = 25, c1t = 1, c2t = 50, c3t = 250,
             varlim = c(0.005, 0.025))
# Constrained optimal design with J = 20
  myod2 <- od.3(icc2 = 0.2, icc3 = 0.1, r12 = 0.5, r22 = 0.5, r32 = 0.5, J = 20,
             c1 = 1, c2 = 5, c3 = 25, c1t = 1, c2t = 50, c3t = 250,
             varlim = c(0, 0.025))
# Relative efficiency (RE)
  myre <- re(od = myod1, subod= myod2)
  myre$re # RE = 0.53

# Unconstrained optimal design of 4-level CRT #---------
  myod1 <- od.4(icc2 = 0.2, icc3 = 0.1, icc4 = 0.05, r12 = 0.5,
              r22 = 0.5, r32 = 0.5, r42 = 0.5,
              c1 = 1, c2 = 5, c3 = 25, c4 = 125,
              c1t = 1, c2t = 50, c3t = 250, c4t = 2500,
              varlim = c(0, 0.01))
# Constrained optimal design with p = 0.5
  myod2 <- od.4(icc2 = 0.2, icc3 = 0.1, icc4 = 0.05, r12 = 0.5, p = 0.5,
              r22 = 0.5, r32 = 0.5, r42 = 0.5,
              c1 = 1, c2 = 5, c3 = 25, c4 = 125,
              c1t = 1, c2t = 50, c3t = 250, c4t = 2500,
              varlim = c(0, 0.01))
# Relative efficiency (RE)
  myre <- re(od = myod1, subod= myod2)
  myre$re # RE = 0.78

}
\references{
(1) Shen, Z., & Kelcey, B. (2020). Optimal sample allocation under
  unequal costs in cluster-randomized trials.
  Journal of Educational and Behavioral Statistics, 45(4): 446–474.
  <https://doi.org/10.3102/1076998620912418>
  (2) Shen, Z., & Kelcey, B. (in press). Optimal sample
  allocation in multisite randomized trials.
  The Journal of Experimental Education.
  <https://doi.org/10.1080/00220973.2020.1830361>
  (3) Shen, Z., & Kelcey, B. (in press).
  Optimal sampling ratios in three-level
  multisite experiments. Journal of Research on Educational Effectiveness.
}
